server/requirements.txt
Django>=2.2.3
django-heroku
gunicorn==19.9.0
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
psycopg2==2.7.3.2
pytz==2017.3
whitenoise==3.3.1